fix: separate nav items gen for install view
