+++ src/components/sidebar/helpers/index.ts
@@ -1,10 +1,12 @@
 import { addNavItemMetaData } from './add-nav-item-meta-data'
+import { generateInstallViewNavItems } from './generate-install-view-nav-items'
 import { generateProductLandingSidebarNavData } from './generate-product-landing-nav-items'
 import { generateTopLevelSidebarNavData } from './generate-top-level-nav-items'
 import { getFilteredNavItems } from './get-filtered-nav-items'
 
 export {
   addNavItemMetaData,
+  generateInstallViewNavItems,
   generateProductLandingSidebarNavData,
   generateTopLevelSidebarNavData,
   getFilteredNavItems,
